+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## TODO - x.x.x
 
-- mu miror in combination with music visualizer
+- [mu miror](https://editor.p5js.org/wyattroy/sketches/s1v8O0WBf) from [youtube](https://www.youtube.com/watch?v=2BCMZMRFljI&t=1s) in combination with music visualizer
 
 ## 0.6.0